third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Intel/resource-timing/test_resource_timing_store_and_clear_during_callback-expected.txt

This is a testharness.js-based test.
PASS No entry should be stored in resource timing buffer since its cleared once an item arrived.
FAIL 6 resource timing entries should be moved to global buffer. assert_equals: 6 resource timing entries should be moved to global buffer. expected 6 but got 7
FAIL http://127.0.0.1:8000/w3c/webperf/resources/empty.js is not expected to be in the Resource Timing buffer assert_unreached: Reached unreachable code
PASS http://127.0.0.1:8000/w3c/resources/testharness.js is expected to have initiatorType script
PASS http://127.0.0.1:8000/w3c/resources/testharnessreport.js is expected to have initiatorType script
PASS http://127.0.0.1:8000/w3c/webperf/resources/empty_script.js is expected to have initiatorType script
PASS http://127.0.0.1:8000/w3c/webperf/resources/resource_timing_test0.js is expected to have initiatorType script
PASS http://127.0.0.1:8000/w3c/webperf/resources/webperftestharness.js is expected to have initiatorType script
PASS http://127.0.0.1:8000/w3c/webperf/resources/webperftestharnessextension.js is expected to have initiatorType script
Harness: the test ran to completion.
